+++ NEWS.md
@@ -14,6 +14,7 @@
 ### New features
 * Added additional input when needed to decide if "Show R Code" should be part of the rendered report.
 * Added support for global `knitr` options in the render method in the Render class.
+* Added the collapsible panel for the Show R Code in the previewer.
 
 ### Breaking changes
 * Simplified `reporter_previewer_ui` to contain just `id` argument.